Split DerivativeBuilder.h based on class
Files created:
- ForwardModeVisitor.cpp
- HessianModeVisitor.cpp
- JacobianModeVisitor.cpp
- ReverseModeVisitor.cpp
- VisitorBase.cpp
